+++ corehq/apps/export/export.py
@@ -27,7 +27,8 @@ from corehq.apps.export.models.new import (
 from corehq.elastic import iter_es_docs_from_query
 from corehq.toggles import PAGINATED_EXPORTS
 from corehq.util.datadog.gauges import datadog_track_errors
-from corehq.util.datadog.utils import DAY_SCALE_TIME_BUCKETS, load_counter
+from corehq.util.datadog.utils import load_counter
+from corehq.util.metrics import DAY_SCALE_TIME_BUCKETS
 from corehq.util.files import TransientTempfile, safe_filename
